LPS-186666 Rename AccessControlledContainerRequestResponseFilter to AccessControlledContainerRequestFilter, so it matches its interface and extract service from AccessControlledContainerRequestFilter to AccessControlledContainerResponseFilter
