+++ modules/apps/portal-remote/portal-remote-jaxrs-security-impl/src/main/java/com/liferay/portal/remote/jaxrs/security/internal/container/response/filter/AccessControlledContainerResponseFilter.java
@@ -1,0 +1,77 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.remote.jaxrs.security.internal.container.response.filter;
+
+import com.liferay.portal.kernel.security.access.control.AccessControlUtil;
+import com.liferay.portal.kernel.security.auth.AccessControlContext;
+
+import java.io.IOException;
+
+import java.util.Map;
+
+import javax.ws.rs.container.ContainerRequestContext;
+import javax.ws.rs.container.ContainerResponseContext;
+import javax.ws.rs.container.ContainerResponseFilter;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.ServiceScope;
+
+/**
+ * @author Joao Victor Alves
+ */
+@Component(
+	property = {
+		"osgi.jaxrs.application.select=(!(liferay.access.control.disable=true))",
+		"osgi.jaxrs.extension=true", "osgi.jaxrs.name=Liferay.Access.Control"
+	},
+	scope = ServiceScope.PROTOTYPE, service = ContainerResponseFilter.class
+)
+public class AccessControlledContainerResponseFilter
+	implements ContainerResponseFilter {
+
+	@Override
+	public void filter(
+			ContainerRequestContext containerRequestContext,
+			ContainerResponseContext containerResponseContext)
+		throws IOException {
+
+		_decrementServiceDepth();
+	}
+
+	private void _decrementServiceDepth() {
+		AccessControlContext accessControlContext =
+			AccessControlUtil.getAccessControlContext();
+
+		if (accessControlContext == null) {
+			return;
+		}
+
+		Map<String, Object> settings = accessControlContext.getSettings();
+
+		Integer serviceDepth = (Integer)settings.get(
+			AccessControlContext.Settings.SERVICE_DEPTH.toString());
+
+		if (serviceDepth == null) {
+			return;
+		}
+
+		serviceDepth--;
+
+		settings.put(
+			AccessControlContext.Settings.SERVICE_DEPTH.toString(),
+			serviceDepth);
+	}
+
+}
